+++ application/common/models/User.php
@@ -24,11 +24,19 @@ class User extends UserBase
     const SCENARIO_AUTHENTICATE    = 'authenticate';
     const SCENARIO_INVITE          = 'invite';
 
+    const NAG_NONE           = 'none';
+    const NAG_ADD_MFA        = 'add_mfa';
+    const NAG_ADD_METHOD     = 'add_method';
+    const NAG_PROFILE_REVIEW = 'profile_review';
+
     /** @var string */
     public $password;
 
     /** @var Ldap */
     private $ldap;
+
+    /** @var string */
+    protected $nagState;
 
     /**
      * {@inheritdoc}
@@ -160,6 +168,8 @@ class User extends UserBase
             'locked',
             'manager_email',
             'require_mfa',
+            'nag_for_mfa_after',
+            'nag_for_method_after',
             'review_profile_after',
             'personal_email',
             'hide',
@@ -205,6 +215,16 @@ class User extends UserBase
             ],
             [
                 'require_mfa', 'default', 'value' => 'no', 'on' => self::SCENARIO_NEW_USER
+            ],
+            [
+                'nag_for_mfa_after',
+                'default',
+                'value' => MySqlDateTime::relative(\Yii::$app->params['mfaAddInterval']),
+            ],
+            [
+                'nag_for_method_after',
+                'default',
+                'value' => MySqlDateTime::relative(\Yii::$app->params['methodAddInterval']),
             ],
             [
                 'review_profile_after',
@@ -497,7 +517,7 @@ class User extends UserBase
                 return $model->getMethodFields();
             },
             'profile_review' => function (self $model) {
-                return $model->isTimeForReview() ? 'yes' : 'no';
+                return $model->getNagState() == self::NAG_PROFILE_REVIEW ? 'yes' : 'no';
             }
         ];
 
@@ -522,13 +542,71 @@ class User extends UserBase
     }
 
     /**
+     * Based on provided time, determine whether to present a reminder to add
+     * an MFA option.
+     * @param int $now
+     * @return bool
+     * @throws Exception
+     */
+    public function isTimeToNagToAddMfa(int $now): bool
+    {
+        return MySqlDateTime::isAfter($now, $this->nag_for_mfa_after)
+            && (count($this->getVerifiedMfaOptions()) === 0);
+    }
+
+    /**
+     * Based on provided time, determine whether to present a reminder to add
+     * a recovery method option.
+     * @param int $now
+     * @return bool
+     * @throws Exception
+     */
+    public function isTimeToNagToAddMethod(int $now): bool
+    {
+        return MySqlDateTime::isAfter($now, $this->nag_for_method_after)
+            && (count($this->getVerifiedMethodOptions()) === 0);
+    }
+
+    /**
      * Based on current time, determine whether to present a profile review to
      * the user.
+     * @param int $now
+     * @return bool
+     * @throws Exception
      */
-    public function isTimeForReview()
+    public function isTimeForReview(int $now)
     {
-        return MySqlDateTime::isBefore($this->review_profile_after, time());
+        return MySqlDateTime::isAfter($now, $this->review_profile_after);
     }
+
+    /**
+     * Based on current time and presence of MFA and Method options,
+     * determine which "nag" to present to the user.
+     *
+     */
+    public function getNagState()
+    {
+        /*
+         * Don't recalculate in case the date has changed since the last calculation.
+         */
+        if ($this->nagState !== null) {
+            return $this->nagState;
+        }
+        $possibleNags = [
+            self::NAG_ADD_MFA => 'isTimeToNagToAddMfa',
+            self::NAG_ADD_METHOD => 'isTimeToNagToAddMethod',
+            self::NAG_PROFILE_REVIEW => 'isTimeForReview',
+        ];
+        $now = time();
+        foreach ($possibleNags as $nagType => $isTime) {
+            if ($this->$isTime($now)) {
+                $this->nagState = $nagType;
+                return $this->nagState;
+            }
+        }
+        return self::NAG_NONE;
+    }
+
 
     /**
      * @return array MFA related properties
@@ -537,6 +615,7 @@ class User extends UserBase
     {
         return [
             'prompt'  => $this->isPromptForMfa() ? 'yes' : 'no',
+            'add'     => $this->getNagState() == self::NAG_ADD_MFA ? 'yes' : 'no',
             'options' => $this->getVerifiedMfaOptions(),
         ];
     }
@@ -586,11 +665,12 @@ class User extends UserBase
      */
     public function getMethodFields()
     {
+        $shouldProvideMethodOptions = $this->getNagState() === self::NAG_PROFILE_REVIEW
+            && $this->scenario == self::SCENARIO_AUTHENTICATE;
+
         return [
-            'options' =>
-                $this->isTimeForReview() && $this->scenario == self::SCENARIO_AUTHENTICATE
-                    ? $this->methods
-                    : [],
+            'add' => $this->getNagState() == self::NAG_ADD_METHOD ? 'yes' : 'no',
+            'options' => $shouldProvideMethodOptions ? $this->methods : [],
         ];
     }
 
@@ -826,11 +906,21 @@ class User extends UserBase
     }
 
     /**
-     * Update the profile review date
+     * Update the date field that corresponds to the current nag state
      */
-    public function updateProfileReviewDate()
+    public function updateProfileReviewDates()
     {
-        $this->review_profile_after = MySqlDateTime::relative(\Yii::$app->params['profileReviewInterval']);
+        switch ($this->getNagState()) {
+            case self::NAG_ADD_MFA:
+                $this->nag_for_mfa_after = MySqlDateTime::relative(\Yii::$app->params['mfaAddInterval']);
+                break;
+            case self::NAG_ADD_METHOD:
+                $this->nag_for_method_after = MySqlDateTime::relative(\Yii::$app->params['methodAddInterval']);
+                break;
+            case self::NAG_PROFILE_REVIEW:
+                $this->review_profile_after = MySqlDateTime::relative(\Yii::$app->params['profileReviewInterval']);
+                break;
+        }
     }
 
     /**
@@ -914,6 +1004,8 @@ class User extends UserBase
             ], 'application');
         }
 
+        $this->setEmptyNagDates();
+
         parent::afterFind();
     }
 
@@ -930,5 +1022,41 @@ class User extends UserBase
         }
 
         return true;
+    }
+
+    /**
+     * Set dates that are empty. These are records that existed before the migration added these
+     * columns in the database. Once all records are updated, this function can be removed.
+     */
+    protected function setEmptyNagDates(): void
+    {
+        $needToSave = false;
+
+        if ($this->review_profile_after === '0000-00-00') {
+            $this->review_profile_after = MySqlDateTime::relative(\Yii::$app->params['profileReviewInterval']);
+            $needToSave = true;
+        }
+
+        if ($this->nag_for_mfa_after === '0000-00-00') {
+            $this->nag_for_mfa_after = MySqlDateTime::relative(\Yii::$app->params['mfaAddInterval']);
+            $needToSave = true;
+        }
+
+        if ($this->nag_for_method_after === '0000-00-00') {
+            $this->nag_for_method_after = MySqlDateTime::relative(\Yii::$app->params['methodAddInterval']);
+            $needToSave = true;
+        }
+
+        if ($needToSave) {
+            $this->scenario = self::SCENARIO_UPDATE_USER;
+            if (! $this->save()) {
+                Yii::warning([
+                    'event' => 'setEmptyNagDates',
+                    'status' => 'save failed',
+                    'employeeId' => $this->employee_id,
+                    'scenario' => $this->scenario,
+                ], 'application');
+            }
+        }
     }
 }